added front end UI

+++ backend/controllers/goalController.js
@@ -1,8 +1,9 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../model/goalModel')
+const User = require('../model/userModel')
 
 const getGoals = asyncHandler(async(req,res)=>{
-    const goals = await Goal.find()
+    const goals = await Goal.find({user:req.user.id})
     res.status(200).json(goals)
 })
 const setGoal = asyncHandler(async(req,res)=>{
@@ -12,7 +13,8 @@ const setGoal = asyncHandler(async(req,res)=>{
         throw new Error("Please add a text field")
     }
     const goal = await Goal.create({
-        text:req.body.text
+        text:req.body.text,
+        user:req.user.id
     })
     res.status(200).json(goal)
 })
@@ -21,6 +23,15 @@ const updateGoal = (asyncHandler(async(req,res)=>{
     if(!goal){
         res.status(400)
         throw new Error('Goal Not Found')
+    } 
+    const user = await User.findById(req.user.id)
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    if(goal.user.toString() !== user.id ){
+        res.status(401)
+        throw new Error('User not authorized')
     } 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id,req.body,{
         new:true
@@ -32,6 +43,15 @@ const deleteGoal = (asyncHandler(async(req,res)=>{
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+    const user = await User.findById(req.user.id)
+    if(!user){
+        res.status(401)
+        throw new Error("User not found")
+    }
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
     await goal.deleteOne()
     res.status(200).json({id:req.params.id})
